Add files via upload
This group project for UAPP 667 course  in MSDS program, University of Delaware was to identify and analyze the areas and factors relating to the high number of road accidents and the resulting injuries and fatalities in India in the years 2016 to 2018, thereafter, predicting the number of road accidents for the year 2019. Both Supervised and Unsupervised machine learning techniques were applied: K-means clustering, Decision tree regressor, and the Random Forest regressor. The methods were evaluated using three measures R squared, root mean squared error, and the mean absolute error.